Get rid of manifest.json logic
