+++ CONTRIBUTING.md
@@ -8,11 +8,6 @@ Please place new tests in one of `tests/asyncapi-2.0` sub-folders. Name of the t
 
 Non-AsyncAPI files (libraries, extensions, etc.) must have a `.yml` extension instead of `.yaml`.
 
-After adding/moving/deleting tck files, manifest has to be regenerated with:
-```sh
-npm run gen-manifest
-```
-
 ## Running tests
 
 We've created a separate project called [tck runner](./runner) to run all the tests contained in the AsyncAPI TCK. By following the instructions on that directory, you should be able to test any new test case that you may want to contribute against the different projects that this runner covers.
