+++ runner/js/package.json
@@ -13,7 +13,8 @@
     "amf-client-js": "^4.0.4",
     "asyncapi-parser": "^0.13.0",
     "minimist": "^1.2.0",
-    "mustache": "^3.0.1"
+    "mustache": "^3.0.1",
+    "walk": "^2.3.14"
   },
   "devDependencies": {}
 }
