+++ runner/js/src/utils.js
@@ -1,11 +1,22 @@
 const path = require('path')
 const fs = require('fs')
+const walk = require('walk')
 
 /* Lists spec files in folder */
-function listFiles (foldPath) {
-  const manifestPath = path.join(foldPath, 'manifest.json')
-  const manifest = JSON.parse(fs.readFileSync(manifestPath, 'utf8'))
-  return manifest.filePaths.map(fp => path.join(foldPath, fp))
+function listYamlFiles (dirPath) {
+  let files = []
+  const options = {
+    listeners: {
+      file: (root, fileStats, next) => {
+        if (fileStats.name.includes('.yaml')) {
+          files.push(path.join(root, fileStats.name))
+        }
+        next()
+      }
+    }
+  }
+  walk.walkSync(dirPath, options)
+  return files.sort()
 }
 
 /* Writes JSON report to output folder */
@@ -17,6 +28,6 @@ function saveReport (report, outdir) {
 }
 
 module.exports = {
-  listFiles: listFiles,
+  listYamlFiles: listYamlFiles,
   saveReport: saveReport
 }
